+++ app/src/main/java/net/pixelos/ota/model/UpdateBaseInfo.java
@@ -43,5 +43,5 @@ public interface UpdateBaseInfo {
 
     String getHash();
 
-    int getMinTimestamp ();
+    long getMinTimestamp ();
 }
